+++ genometryImpl/src/com/affymetrix/genometryImpl/symmetry/UcscBedSym.java
@@ -214,11 +214,11 @@ public class UcscBedSym implements SeqSpan, SupportsCdsSpan, TypedSym, SymWithPr
 	public int getLength() { return (txMax - txMin); }
 	public boolean isForward() { return forward; }
 	public BioSeq getBioSeq() { return seq; }
-	public double getStartDouble() { return (double)getStart(); }
-	public double getEndDouble() { return (double)getEnd(); }
-	public double getMaxDouble() { return (double)getMax(); }
-	public double getMinDouble() { return (double)getMin(); }
-	public double getLengthDouble() { return (double)getLength(); }
+	public double getStartDouble() { return getStart(); }
+	public double getEndDouble() { return getEnd(); }
+	public double getMaxDouble() { return getMax(); }
+	public double getMinDouble() { return getMin(); }
+	public double getLengthDouble() { return getLength(); }
 	public boolean isIntegral() { return true; }
 
 	public float getScore() { return score; }
